+++ packages/components/vite.build.config.ts
@@ -1,11 +1,11 @@
-import banner from "vite-plugin-banner";
-import dts from "vite-plugin-dts";
+import dts from "unplugin-dts/vite";
 import baseConfig from "./vite.config";
 import { externalizeDeps } from "vite-plugin-externalize-deps";
 import { defineConfig, mergeConfig } from "vite";
-import { flowComponentsLayerPlugin } from "./dev/vite/flowComponentsLayerPlugin";
-import { layerOrderPlugin } from "./dev/vite/layerOrderPlugin";
-import { stylesheetVariantsPlugin } from "./dev/vite/stylesheetVariantsPlugin";
+import { flowComponentsLayerPlugin } from "./dev/vite/flowComponentsLayerPlugin.ts";
+import { layerOrderPlugin } from "./dev/vite/layerOrderPlugin.ts";
+import { stylesheetVariantsPlugin } from "./dev/vite/stylesheetVariantsPlugin.ts";
+import { preserveUseClientBanner } from "../core";
 
 export default mergeConfig(
   baseConfig,
@@ -31,8 +31,9 @@ export default mergeConfig(
         formats: ["es"],
       },
       emptyOutDir: false,
-      rollupOptions: {
+      rolldownOptions: {
         output: {
+          postBanner: preserveUseClientBanner,
           format: "es",
           preserveModules: true,
           entryFileNames: "js/[name].mjs",
@@ -61,11 +62,6 @@ export default mergeConfig(
     plugins: [
       layerOrderPlugin(),
       stylesheetVariantsPlugin(),
-      banner((filename) =>
-        filename.endsWith(".mjs") && !filename.endsWith("index.mjs")
-          ? '"use client"\r\n/* */'
-          : "",
-      ),
       externalizeDeps({
         except: ["@mittwald/flow-design-tokens/**/*", "@mittwald/flow-core"],
       }),